editor/composer/test/test_bug348497.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=348497
-->
<head>
  <title>Test for Bug 348497</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=348497">Mozilla Bug 348497</a>
<p id="display"></p>
<div id="content">
  This page should not crash Mozilla<br>
  <iframe id="testIframe"></iframe>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 348497 **/
function doe() {
    document.getElementById('testIframe').style.display = 'block';
    document.getElementById('testIframe').contentDocument.designMode = 'on';
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doe);
addLoadEvent(function() { ok(true, "enabling designmode on an iframe onload does not crash Mozilla")});
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
